+++ lib/bindings/binding.dart
@@ -1,10 +1,10 @@
 part of fmvvm.bindings;
 
 /// Used for a binding between a widget and a object that implements BindableBase.
-/// 
+///
 /// This item is usually created within a FmvvmState object.
 class Binding {
-  Binding(BindableBase source, PropertyInfo sourceProperty, 
+  Binding(BindableBase source, PropertyInfo sourceProperty,
       {BindingDirection bindingDirection,
       fmvvm_interfaces.ValueConverter valueConverter}) {
     this.source = source;
@@ -20,12 +20,12 @@ class Binding {
 
   fmvvm_interfaces.ValueConverter _valueConverter;
 
-  /// The source bindable base object.  
+  /// The source bindable base object.
   BindableBase source;
 
   /// If the binding only happens once or if it is able to be bi-directional.
   BindingDirection get bindingDirection => _bindingDirection;
-  
+
   /// The propertyInfo object being bound to on the source.
   PropertyInfo get sourceProperty => _sourceProperty;