+++ lib/core.dart
@@ -6,9 +6,9 @@ class Core {
   static NavigationService _navigationService;
   static ViewLocator _viewLocator;
 
-  /// Called to initialize the fmvvm framework. 
-  /// 
-  /// This should be done before attempting to use any fmvvm components, 
+  /// Called to initialize the fmvvm framework.
+  ///
+  /// This should be done before attempting to use any fmvvm components,
   /// usually during the app startup.
   /// The [registrations] parameter allows overrides of the devault navigation,
   /// view location and component resolution (dependency injection/Ioc) implementations.
@@ -19,12 +19,13 @@ class Core {
     _viewLocator = startRegistrations.getViewLocator();
     _componentResolver = startRegistrations.getResolver();
 
-      _componentResolver.registerInstance<ViewLocator>(_viewLocator);
-      _componentResolver.registerInstance<NavigationService>(_navigationService);
+    _componentResolver.registerInstance<ViewLocator>(_viewLocator);
+    _componentResolver.registerInstance<NavigationService>(_navigationService);
   }
 
   /// A global reference to the registered object for dependency injection/IoC.
   static ComponentResolver get componentResolver => _componentResolver;
+
   /// A global reference to the navigation service used by the system.
-  /// 
+  ///
 }