+++ CHANGELOG.md
@@ -1,6 +1,15 @@
+## [0.8.1] - 11/6/2019.
+
+* This release adds:
+ 
+- notificationList for binding to lists
+- Allowed Stateful and Stateless widgets to bind to and BindableBase object, not just viewModels
+- Clean up project
+- Add list samples
+
 ## [0.8.0] - 11/5/2019.
 
-* TODO: Initial release includes:
+* Initial release includes:
  
 - Dependency injection
 - Data binding